@@ -10,16 +10,9 @@
     "ng": "ng",
     "serve": "ng serve",
     "start": "node server.js",
-    "start:local": "ng serve --environment=local",
-    "start:dev:remote": "ng serve --environment=dev_remote",
     "start:dev:local": "ng serve --environment=dev_local",
-    "start:staging:remote": "ng serve --aot --environment=staging_remote",
-    "start:staging:local": "ng serve --aot --environment=staging_local",
-    "start:prod:remote": "ng serve --aot --prod --environment=prod_remote",
-    "start:prod:local": "ng serve --aot --prod --environment=prod_local",
+    "start:dev:heroku": "ng serve --environment=dev_local_heroku",
     "build": "ng build",
-    "build:prod:local": "ng build --prod --aot --environment=prod_local",
-    "build:prod:heroku": "ng build --prod --aot --environment=prod_remote",
     "test": "ng test",
     "test_single_run": "ng test --single-run",
     "lint": "ng lint",
@@ -53,9 +46,9 @@
     "materialize-css": "^0.97.8",
     "mdi": "^1.8.36",
     "ngx-pagination": "^3.0.1",
-    "rxjs": "^5.1.0",
+    "rxjs": "^5.4.2",
     "ts-node": "~2.0.0",
-    "typescript": "~2.2.2",
+    "typescript": "~2.5.3",
     "underscore": "^1.8.3",
     "zone.js": "^0.7.6"
   },
