@@ -1,10 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Http, Headers, Response, RequestOptions, URLSearchParams, Jsonp } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
-import { environment } from '../../../environments/environment';
-import { Organization } from './organization';
-import { AuthHttp } from 'angular2-jwt';
+import {Injectable} from '@angular/core';
+import {Http, Headers, Response, RequestOptions, URLSearchParams, Jsonp} from '@angular/http';
+import {Observable} from 'rxjs/Observable';
+import {Subject} from 'rxjs/Subject';
+import {environment} from '../../../environments/environment';
+import {Organization} from './organization';
+import {AuthHttp} from 'angular2-jwt';
 
 const organizationUrl = `${environment.backend_url}/api/organizations`;
 
@@ -13,42 +13,41 @@ export class OrganizationService {
   private organizationLinkedSource = new Subject<string>();
   public organizationLinkedSource$ = this.organizationLinkedSource.asObservable();
 
-  constructor(private http: Http, private jsonp: Jsonp, private authHttp: AuthHttp) { }
+  constructor(private http: Http, private jsonp: Jsonp, private authHttp: AuthHttp) {
+  }
 
   getOrganizations(): Observable<Organization[]> {
     return this.http
-               .get(organizationUrl)
-               .map(res => res.json());
+    .get(organizationUrl)
+    .map(res => res.json());
   }
 
   getOrganization(id: number): Observable<Organization> {
     return this.http
-               .get(`${organizationUrl}/${id}`)
-               .map(res => res.json());
+    .get(`${organizationUrl}/${id}`)
+    .map(res => res.json());
   }
 
   getUserOrganization(id: number): Observable<Organization[]> {
     return this.http
-               .get(`${organizationUrl}/user/${id}`)
-               .map(res => res.json());
+    .get(`${organizationUrl}/user/${id}`)
+    .map(res => res.json());
   }
 
   approve(organizationId: number, status: string) {
     const requestOptions = new RequestOptions();
     requestOptions.search = new URLSearchParams(`status=${status}`);
     return this.authHttp
-      .put(`${organizationUrl}/${organizationId}/approve`, '', requestOptions);
+    .put(`${organizationUrl}/${organizationId}/approve`, '', requestOptions);
   }
 
-  searchOrganizations(
-    keyword?: string,
-    countries?: string[],
-    open?: boolean,
-    status?: string,
-    category?: string[],
-    page?: number,
-    size?: number
-  ): Observable<any> {
+  searchOrganizations(keyword?: string,
+                      countries?: string[],
+                      open?: boolean,
+                      status?: string,
+                      category?: string[],
+                      page?: number,
+                      size?: number): Observable<any> {
     const params = new URLSearchParams();
 
     if (keyword) {
@@ -84,16 +83,16 @@ export class OrganizationService {
     }
 
     return this.http
-      .get(`${organizationUrl}/search`, {search: params})
-      .map( res => ({data: res.json().content, totalItems: res.json().totalElements}))
-      .catch(this.handleError);
+    .get(`${organizationUrl}/search`, {search: params})
+    .map(res => ({data: res.json().content, totalItems: res.json().totalElements}))
+    .catch(this.handleError);
   }
 
-  createOrganization(organization: Organization): Observable<{organization: Organization}> {
-      return this.authHttp.post(
+  createOrganization(organization: Organization): Observable<{ organization: Organization }> {
+    return this.authHttp.post(
       `${organizationUrl}`,
       organization
-      ).map(res => res.json());
+    ).map(res => res.json());
   }
 
   linkUserOrganization(userId: String, organizationId: number) {
@@ -110,10 +109,10 @@ export class OrganizationService {
   }
 
   updateOrganization(organization: Organization): Observable<Response> {
-      return this.authHttp.put(
+    return this.authHttp.put(
       `${organizationUrl}/${organization.id}`,
       organization
-      );
+    );
   }
 
   delete(id: number): Observable<Response> {
@@ -127,7 +126,13 @@ export class OrganizationService {
     const requestOptions = new RequestOptions();
     requestOptions.search = new URLSearchParams(`imgUrl=${imgUrl}`);
     return this.authHttp
-      .put(`${organizationUrl}/${id}/logo`, '', requestOptions);
+    .put(`${organizationUrl}/${id}/logo`, '', requestOptions);
+  }
+
+  getTotalCountries(): Observable<any> {
+    return this.http
+    .get(`${organizationUrl}/countries/total`)
+    .map(res => res.json());
   }
 
   private handleError(error: any): Promise<any> {
