@@ -3,12 +3,14 @@ export class Applicant {
 constructor (
     public userId: number,
     public projectId: number,
+    public applicationStatus: string,
+    public resumeFlag: boolean,
     public firstName?: string,
     public lastName?: string,
     public title?: string,
-    public applicationStatus?: string,
-    public appliedTime?: string,
-    public acceptedTime?: string,
-    public declined?: string
+    public comment?: string,
+    public appliedTime?: Date,
+    public acceptedTime?: Date,
+    public declinedTime?: Date
   ) { }
 }
