@@ -1,4 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Location } from '@angular/common';
+import { Hero } from '../user/common/hero';
+import { ProjectService } from '../project/common/project.service';
 
 @Component({
   selector: 'my-appreciations',
@@ -7,9 +10,25 @@ import { Component, OnInit } from '@angular/core';
 })
 export class AppreciationsComponent implements OnInit {
 
-  constructor(
-  ) { }
+  heroes: Hero[];
+  rowsArray:  number[];
+
+  constructor(	private projectService: ProjectService,
+    private location: Location) {
+  }
 
   ngOnInit(): void {
+    this.projectService.getHeroes()
+        .subscribe(
+          res =>	{
+          this.heroes = res;
+          this.rowsArray = Array.from(Array(Math.ceil(this.heroes.length / 3)).keys());
+          },
+          error =>	console.log(error)
+        );
+  }
+
+  goBack(): void {
+    this.location.back();
   }
 }
