@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Router,
-         ActivatedRouteSnapshot,
-         RouterStateSnapshot } from '@angular/router';
+import { Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { CanActivate } from '@angular/router';
 import { AuthService } from './auth.service';
 import { Location } from '@angular/common';
@@ -13,6 +11,7 @@ export class AuthGuard implements CanActivate {
 
   canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     const appMetaData = 'http://app_metadata';
+
     if (this.auth.authenticated()) {
         const roles = next.data['roles'] as Array<string>;
         // Check if roles have NOT been set along with [AuthGuard] marker in auth.routing.ts
