@@ -1,12 +1,12 @@
 export const environment = {
     production: false,
-    backend_url: 'https://c4sg-dev.herokuapp.com',
+    backend_url: 'http://localhost:8080',
     auth_clientID: 'oO8PRySIJnGHDGsZQ36zOhhccNE6iNjf',
     auth_domain: 'c4sg-dev.auth0.com',
     auth_tenant_shared: false,
     auth_api: 'https://c4sg-api',
     auth_callback_env: 'local',
-    auth_silenturl: 'silent-local.html',
+    auth_silenturl: 'silent-dev-local.html',
     AWS_IMAGE_BUCKET: 'c4sg.dev.images',
     AWS_DOCS_BUCKET: 'c4sg.dev.docs',
     AWS_REGION: 'us-west-2',
