@@ -1,12 +1,12 @@
 export const environment = {
     production: true,
-    backend_url: 'https://c4sg-staging.herokuapp.com',
-    auth_clientID: 'xwMK322G3bOF4OV20tqDlaMx2TMYjc32',
+    backend_url: 'https://c4sg-staging.herokuapp.com/',
+    auth_clientID: 'PzdL8OyL4MYgFOGZAmQ1Qog8dXkbliil',
     auth_domain: 'c4sg-staging.auth0.com',
     auth_tenant_shared: true,
     auth_api: 'https://c4sg-api',
     auth_callback_env: 'staging',
-    auth_silenturl: 'silent-staging.html',
+    auth_silenturl: 'silent-staging-remote.html',
     AWS_IMAGE_BUCKET: 'c4sg.dev.images',
     AWS_DOCS_BUCKET: 'c4sg.dev.docs',
     AWS_REGION: 'us-west-2',
