@@ -6,12 +6,12 @@
 export const environment = {
     production: false,
     backend_url: 'http://localhost:8080',
-    auth_clientID: 'oO8PRySIJnGHDGsZQ36zOhhccNE6iNjf',
-    auth_domain: 'c4sg-dev.auth0.com',
+    auth_clientID: 'EIefKrM2FEELy4E66NJ9uJ4QPvjU0LOO',
+    auth_domain: 'c4sg-local.auth0.com',
     auth_tenant_shared: false,
     auth_api: 'https://c4sg-api',
-    auth_callback_env: 'dev',
-    auth_silenturl: 'silent-dev.html',
+    auth_callback_env: 'local',
+    auth_silenturl: 'silent-dev-local.html',
     AWS_IMAGE_BUCKET: 'c4sg.dev.images',
     AWS_DOCS_BUCKET: 'c4sg.dev.docs',
     AWS_REGION: 'us-west-2',
